Rework the NU1510 heuristic, only raise it when a package id can be completely removed from the project
